Simplify repositories.
