--- app/Core/Repository/Contracts/Activity/CanBeCreated.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace FELS\Core\Repository\Contracts\Activity;
-
-interface CanBeCreated
-{
-    /**
-     * Create a new model instance.
-     *
-     * @param array $data
-     * @return static
-     */
-    public function create(array $data);
-}
