--- app/Core/Repository/Contracts/Activity/CanBeRemoved.php
@@ -1,30 +1,0 @@
-<?php
-
-namespace FELS\Core\Repository\Contracts\Activity;
-
-interface CanBeRemoved
-{
-    /**
-     * Restore a soft deleted model instance.
-     *
-     * @param $identifier
-     * @return bool|null
-     */
-    public function restore($identifier);
-
-    /**
-     * Soft delete a model instance.
-     *
-     * @param $identifier
-     * @return bool|null
-     */
-    public function softDelete($identifier);
-
-    /**
-     * Permanently delete a soft deleted model instance.
-     *
-     * @param $identifier
-     * @return void
-     */
-    public function forceDelete($identifier);
-}
