--- app/Core/Repository/Contracts/Activity/CanBeUpdated.php
@@ -1,15 +1,0 @@
-<?php
-
-namespace FELS\Core\Repository\Contracts\Activity;
-
-interface CanBeUpdated
-{
-    /**
-     * Update a model instance.
-     *
-     * @param array $data
-     * @param $identifier
-     * @return bool|int
-     */
-    public function update(array $data, $identifier);
-}
